+++ backend-v3/CHANGELOG.md
@@ -19,6 +19,7 @@ After version 3.1.0, this project follow these guidelines,
 ## [3.11.2] - 2020-03-06
 ### Added
   - `kumabank` settlement on agent level
+  - `kumabank` void transaction
 ### Removed
   - `kumabank` random failure in creating transaction